Manufacturer searcher now functional
Nearly there. Correctly displays and builds up JQL when selecting, but
does not invalidate or rebuild the manufacturer list when selecting a
single catalog that has a specific manufacturer context.

Basically, if a catalog only has one manufacturer associated with it,
the manufacturer list should only show that one manufacturer and display
that the previously selected one are now invalid and should be
deselected.