Search works! :gun: :boom:
So it turns out that the Lucene writerReference commit was actually the
culprit for the deleted indexes. I believe this is related to Lucene's
default commit policy which keeps only the commit that is currently
being operated on, and since the commit we were doing was empty, it felt
that it should delete all the segments. I think.

Getting back to work on the asset navigator itself. Need to hook
searcher updates into the layout re-render calls, which for some reason
is not working right now.

TODO: Include the SKU into the asset index, and work on getting the
asset table styled correctly. Also, look into caching manufacturer image
thumbnails so that they can be displayed next to each asset table
result.