The asset table now nearly works
Updating a searcher will now trigger a re-render of the split-view
layout.

Right now the asset table is rendered from the server, then wiped out
due to the fact that the search results are not being stored correctly
in the search module. Once this get's fixed and the layouts unify, the
user should not be able to tell that there is some server side view
magic going on here.

Started work on the asset viewer code. Basically, I am going to
implement a custom "AppModule" which will assign a set of module
requests/commands so that Wreqr can do it's thing. This will allow all
those trigger("assetEditor:*") calls to work.