Asset header is now visible on asset click
The asset now loads on click, and renders the header bits so far. I have
a disposal issue after clicking a second asset, but that will most
likely be a simple fix.

Still lots of work to do as far as rendering panels go, but I might just
fudge that and see what I can work around in the short term.